+++ spec/examples/inherit_from_increment_value.rb
@@ -1,8 +1,6 @@
 # frozen_string_literal: true
 
 class InheritFromIncrementValue < IncrementValue
-  input :name, type: 'String'
-
   def call
     super