+++ spec/examples/use_unknown_input.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 class UseUnknownInput < Actor
-  input :name
+  input :name, default: 'Jim'
 
   def call
     context.foobar